Implement Schema object
The Schema object allows classes created from sub-schemas to be extended when
their parent class' schema is extended. It also relaxes the constraint of
defining schemas with a type right away. Instead, a true base class can be made,
then extended from to create subclasses that can represent arrays and objects
(both mutable and immutable) instead of being forced to choose Array,
ImmutableObject, or MutableObject to extend from the beginning.

The Schema object provides a foundation for schema-related enhancements in the
future, such as validation of schemas during class definition instead of invalid
schemas causing errors later on.